Fixes Slideshow
git-svn-id: https://buemrmfls01/svn/SITE2017/LOREAL/LORA254-Loreal-TV@2815 bafdefb5-79ff-3e4c-8193-be25f676ad87

+++ Source/app/Http/Controllers/MediaController.php
@@ -191,7 +191,7 @@ class MediaController extends Controller
         $return = true;
         if($data){
             foreach($data as $slide){
-                if(isset($slide['willDelete'])){
+                if(isset($slide['willDelete']) && isset($slide['id'])){
                     $this->deleteSlide($slide);
                 } elseif($slide['video_loop'] != 0 && $slide['video_loop'] != ''){
                     $return = $this->slideVideo($slide);
